- Ability to write configuration into memory - CLean up: Separate models and interfaces
