--- src/vs/platform/configuration/common/model.ts
@@ -1,194 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-'use strict';
-
-import { Registry } from 'vs/platform/registry/common/platform';
-import * as json from 'vs/base/common/json';
-import { IConfigurationRegistry, Extensions, OVERRIDE_PROPERTY_PATTERN } from 'vs/platform/configuration/common/configurationRegistry';
-import { ConfigurationModel, IOverrides, overrideIdentifierFromKey } from 'vs/platform/configuration/common/configuration';
-
-export function getDefaultValues(): any {
-	const valueTreeRoot: any = Object.create(null);
-	const properties = Registry.as<IConfigurationRegistry>(Extensions.Configuration).getConfigurationProperties();
-
-	for (let key in properties) {
-		let value = properties[key].default;
-		addToValueTree(valueTreeRoot, key, value, message => console.error(`Conflict in default settings: ${message}`));
-	}
-
-	return valueTreeRoot;
-}
-
-export function toValuesTree(properties: { [qualifiedKey: string]: any }, conflictReporter: (message: string) => void): any {
-	const root = Object.create(null);
-
-	for (let key in properties) {
-		addToValueTree(root, key, properties[key], conflictReporter);
-	}
-
-	return root;
-}
-
-function addToValueTree(settingsTreeRoot: any, key: string, value: any, conflictReporter: (message: string) => void): void {
-	const segments = key.split('.');
-	const last = segments.pop();
-
-	let curr = settingsTreeRoot;
-	for (let i = 0; i < segments.length; i++) {
-		let s = segments[i];
-		let obj = curr[s];
-		switch (typeof obj) {
-			case 'undefined':
-				obj = curr[s] = Object.create(null);
-				break;
-			case 'object':
-				break;
-			default:
-				conflictReporter(`Ignoring ${key} as ${segments.slice(0, i + 1).join('.')} is ${JSON.stringify(obj)}`);
-				return;
-		}
-		curr = obj;
-	};
-
-	if (typeof curr === 'object') {
-		curr[last] = value; // workaround https://github.com/Microsoft/vscode/issues/13606
-	} else {
-		conflictReporter(`Ignoring ${key} as ${segments.join('.')} is ${JSON.stringify(curr)}`);
-	}
-}
-
-export function getConfigurationKeys(): string[] {
-	const properties = Registry.as<IConfigurationRegistry>(Extensions.Configuration).getConfigurationProperties();
-
-	return Object.keys(properties);
-}
-
-export class DefaultConfigurationModel<T> extends ConfigurationModel<T> {
-
-	constructor() {
-		super(getDefaultValues());
-		this._keys = getConfigurationKeys();
-		this._overrides = Object.keys(this._contents)
-			.filter(key => OVERRIDE_PROPERTY_PATTERN.test(key))
-			.map(key => {
-				return <IOverrides<any>>{
-					identifiers: [overrideIdentifierFromKey(key).trim()],
-					contents: toValuesTree(this._contents[key], message => console.error(`Conflict in default settings file: ${message}`))
-				};
-			});
-	}
-
-	public get keys(): string[] {
-		return this._keys;
-	}
-}
-
-interface Overrides<T> extends IOverrides<T> {
-	raw: any;
-}
-
-export class CustomConfigurationModel<T> extends ConfigurationModel<T> {
-
-	protected _parseErrors: any[] = [];
-
-	constructor(content: string = '', private name: string = '') {
-		super();
-		if (content) {
-			this.update(content);
-		}
-	}
-
-	public get errors(): any[] {
-		return this._parseErrors;
-	}
-
-	public update(content: string): void {
-		let parsed: T = <T>{};
-		let overrides: Overrides<T>[] = [];
-		let currentProperty: string = null;
-		let currentParent: any = [];
-		let previousParents: any[] = [];
-		let parseErrors: json.ParseError[] = [];
-
-		function onValue(value: any) {
-			if (Array.isArray(currentParent)) {
-				(<any[]>currentParent).push(value);
-			} else if (currentProperty) {
-				currentParent[currentProperty] = value;
-			}
-			if (OVERRIDE_PROPERTY_PATTERN.test(currentProperty)) {
-				onOverrideSettingsValue(currentProperty, value);
-			}
-		}
-
-		function onOverrideSettingsValue(property: string, value: any): void {
-			overrides.push({
-				identifiers: [overrideIdentifierFromKey(property).trim()],
-				raw: value,
-				contents: null
-			});
-		}
-
-		let visitor: json.JSONVisitor = {
-			onObjectBegin: () => {
-				let object = {};
-				onValue(object);
-				previousParents.push(currentParent);
-				currentParent = object;
-				currentProperty = null;
-			},
-			onObjectProperty: (name: string) => {
-				currentProperty = name;
-			},
-			onObjectEnd: () => {
-				currentParent = previousParents.pop();
-			},
-			onArrayBegin: () => {
-				let array: any[] = [];
-				onValue(array);
-				previousParents.push(currentParent);
-				currentParent = array;
-				currentProperty = null;
-			},
-			onArrayEnd: () => {
-				currentParent = previousParents.pop();
-			},
-			onLiteralValue: onValue,
-			onError: (error: json.ParseErrorCode) => {
-				parseErrors.push({ error: error });
-			}
-		};
-		if (content) {
-			try {
-				json.visit(content, visitor);
-				parsed = currentParent[0] || {};
-			} catch (e) {
-				console.error(`Error while parsing settings file ${this.name}: ${e}`);
-				this._parseErrors = [e];
-			}
-		}
-		this.processRaw(parsed);
-
-		const configurationProperties = Registry.as<IConfigurationRegistry>(Extensions.Configuration).getConfigurationProperties();
-		this._overrides = overrides.map<IOverrides<T>>(override => {
-			// Filter unknown and non-overridable properties
-			const raw = {};
-			for (const key in override.raw) {
-				if (configurationProperties[key] && configurationProperties[key].overridable) {
-					raw[key] = override.raw[key];
-				}
-			}
-			return {
-				identifiers: override.identifiers,
-				contents: <T>toValuesTree(raw, message => console.error(`Conflict in settings file ${this.name}: ${message}`))
-			};
-		});
-	}
-
-	protected processRaw(raw: T): void {
-		this._contents = toValuesTree(raw, message => console.error(`Conflict in settings file ${this.name}: ${message}`));
-		this._keys = Object.keys(raw);
-	}
-}
