+++ src/vs/platform/configuration/test/common/configurationModel.test.ts
@@ -5,7 +5,7 @@
 'use strict';
 
 import * as assert from 'assert';
-import { CustomConfigurationModel, DefaultConfigurationModel } from 'vs/platform/configuration/common/model';
+import { CustomConfigurationModel, DefaultConfigurationModel } from 'vs/platform/configuration/common/configurationModels';
 import { Extensions, IConfigurationRegistry } from 'vs/platform/configuration/common/configurationRegistry';
 import { Registry } from 'vs/platform/registry/common/platform';
